+++ tests/testthat/test-9-dynamic.R
@@ -1734,6 +1734,17 @@ test_with_dir("no dynamic file_out() (#1141)", {
   expect_error(make(plan), regexp = "file_out")
 })
 
+test_with_dir("no dynamic knitr_in() (#1229)", {
+  plan <- drake_plan(
+    index = c(1L, 2L, 3L, 4L),
+    write_files = target(
+      do_stuff(index, report = knitr_in("report.Rmd")),
+      dynamic = map(index)
+    )
+  )
+  expect_error(suppressWarnings(make(plan)), regexp = "knitr_in")
+})
+
 test_with_dir("log dynamic target as failed if a sub-target fails (#1158)", {
   plan <- drake_plan(
     x = seq_len(2),