+++ NEWS.md
@@ -34,6 +34,8 @@
 * Deprecate `hasty_build` (#1222).
 * Migrate constant runtime parameters to `config$settings` (#965).
 * Warn the user if `file_in()`/`file_out()`/`knitr_in()` files are not literal strings (#1229).
+* Prohibit `file_out()` and `knitr_in()` in imported functions (#1229).
+* Prohibit `knitr_in()` in dynamic branching (#1229).
 
 
 # Version 7.11.0